Added and then removed Disqus to cli
Seems to be impossible to get the user's username due to XSS attacks and the like. Gosh darned security, forcing me to find a vulnerability for a gimmick feature, oof. At any rate, because of this, I don't think it's possible to have the Disqus username of the person autofill in the CLI, unfortunately. I can instead make it fill for their website login, which should be a lot more doable and not require the discovery and usage of some vulnerability in the current XSS prevention methods.